ForceAnalysis.jl

ForceAnalysis.ForceData

ForceAnalysis.ForceEpochs

ForceAnalysis.ForceResponse

ForceAnalysis.MultiForceData

ForceAnalysis.OnsetCriterion

Base.copy

Base.maximum

Base.minimum

DataFrames.aggregate

DataFrames.subset

FileIO.load

FileIO.save

ForceAnalysis.adjust_baseline!

ForceAnalysis.epoch_rejection

ForceAnalysis.epochs

ForceAnalysis.force

ForceAnalysis.impulse_size

ForceAnalysis.latency

ForceAnalysis.lowpass_filter!

ForceAnalysis.response_detection

ForceAnalysis.scale_force!

Statistics.mean

Statistics.median

Statistics.std

Statistics.var

ForceAnalysis.ForceData — Type

ForceData{T <: AbstractFloat}

TODO

ForceAnalysis.ForceEpochs — Type

ForceEpochs{T <: AbstractFloat}

TODO

ForceAnalysis.ForceResponse — Type

ForceResponse

TODO

ForceAnalysis.MultiForceData — Type

MultiForceData{T <: AbstractFloat}

TODO

ForceAnalysis.OnsetCriterion — Type

OnsetCriterion{T <: AbstractFloat}

TODO

Base.copy — Method

copy(fe::ForceEpochs)

TODO

Base.maximum — Method

maximum(fe:ForceEpochs)

Minimum of each epoch.

Base.minimum — Method

minimum(fe:ForceEpochs)

Minimum of each epoch.

DataFrames.aggregate — Method

aggregate(fe::ForceEpochs; condition::ColumnIndex = :all,
	subject_id::Union{Nothing, ColumnIndex} = nothing, agg_fnc::Function = mean)

TODO

DataFrames.subset — Method

subset(fe::ForceEpochs, rows::Base.AbstractVecOrTuple{Integer})
subset(fe::ForceEpochs, args...)

TODO

FileIO.load — Method

load(::Type{ForceData}, filename::String)
load(::Type{ForceEpochs}, [format::Symbol], filename::String)

Loads ForceData or ForceEpochs as jdl2 file.

If the format of the ForceEpochs is not specified, it will be infered from the filename suffix.

FileIO.save — Method

save(filename::String, fd::ForceData; compress = true)
save(filename::String, [format::Symbol], fe::ForceEpochs; compress = true)

Saves ForceData or ForceEpochs as jld2 file.

For ForceEpochs the format can be optionally specified:

:jld2: jld2 file

:csv: creates a Zip-file containing seperate csv-data for forces, design,

 `baseline` and a json file with `sampling_rate` and `zero_times`

ForceAnalysis.adjust_baseline! — Method

adjust_baseline!(fe::ForceEpochs, baseline_window::UnitRange{<:Integer})

TODO

ForceAnalysis.epoch_rejection — Method

epoch_rejection(fe::ForceEpochs; force_range::UnitRange, max_difference::Integer,
	max_diff_windows_size::Integer)

TODO

ForceAnalysis.epochs — Method

epochs(fd::ForceData{<: AbstractFloat};
	zero_times::AbstractVector{<:Integer},
	n_samples::Integer,	n_samples_before::Integer)

TODO

ForceAnalysis.force — Method

force(fd::ForceData)
force(fd::MultiForceData, id::Union{Int, String, Symbol})
force(fe::ForceEpochs)

Returns the force data as Matrix or Vector.

ForceAnalysis.impulse_size — Method

impulse_size(force_vector::AbstractVector{<:AbstractFloat}, rb::ForceResponse)
impulse_size(fp::ForceEpochs, rb::AbstractVector{ForceResponse})

TODO

ForceAnalysis.latency — Method

latency(force_vector::AbstractVector{<:AbstractFloat}, rb::ForceResponse)
latency(fp::ForceEpochs, rb::AbstractVector{ForceResponse})

TODO

ForceAnalysis.lowpass_filter! — Method

lowpass_filter!(fd::ForceData; cutoff_freq::Real,butterworth_order::Integer = 4)
lowpass_filter!(fd::MultiForceData; cutoff_freq::Real,butterworth_order::Integer = 4)
lowpass_filter!(fd::ForceEpochs; cutoff_freq::Real,butterworth_order::Integer = 4)

TODO

ForceAnalysis.response_detection — Method

response_detection(force_vector::AbstractVector{<:AbstractFloat}, criterion::OnsetCriterion; zero_sample::Integer)::ForceResponse
response_detection(fe::ForceEpochs, criterion::OnsetCriterion)::Vector{ForceResponse}

Returns a ForceResponse or Vector{ForceResponse} TODO

ForceAnalysis.scale_force! — Method

scale_force!(fd::ForceData, factor::Real)
scale_force!(fe::ForceEpochs, factor::Real)

TODO

Statistics.mean — Method

mean(fe::ForceEpochs; rows::Union{Nothing, BitVector, Vector{<:Integer}}=nothing)

Computes the mean of the epochs, i.e., column-wise means for each sample. If rows is defined, only the selected rows are considered.

Statistics.median — Method

median(fe::ForceEpochs; rows::Union{Nothing, BitVector, Vector{<:Integer}}=nothing)

Computes the median of the epochs, i.e., column-wise medians for each sample. If rows is defined, only the selected rows are considered.

Statistics.std — Method

std(fe::ForceEpochs; rows::Union{Nothing, BitVector, Vector{<:Integer}}=nothing)

Computes the standard deviation of the epochs, i.e., column-wise standard deviations for each sample. If rows is defined, only the selected rows are considered.

Statistics.var — Method

var(fe::ForceEpochs; rows::Union{Nothing, BitVector, Vector{<:Integer}}=nothing)

Computes the variance of the epochs, i.e., column-wise variances for each sample. If rows is defined, only the selected rows are considered.